requirements.txt
fdt>=0.1.0
imx>=0.1.2
click>=7.0
uboot>=0.1.0
PyYAML>=3.10
Jinja2>=2.6
easy-enum>=0.1.0
voluptuous